vimrc: add per-project exclude file for TagHL
There is no option in TagHL to exclude files or directories, but ctags
has an option to exclude patterns from a file, so just tell TagHL to
read an exclude file in the current directory (if the file doesn't
exist, ctags just ignores the flag).

Also stop ignoring override, as universal-ctags has a new shiny CXX
parser that is C++11-aware.