vimrc: set number globally
Setting number in a FileType autocmd is broken, because it is a
window-local option, meaning that if the same buffer is opened in
another window, the autocmd will not be triggered (even though we want
it to).

Since it doesn't really hurt to always display line numbers (and avoids
having to enable it for every single language), let's just set number
globally.