vimrc: fix colorcolumn setting and cleanup autocmds
colorcolumn is window-local, so setting it in a FileType autocmd is not
good enough. Work around that by setting a buffer-local variable
instead, and checking it in a BufWinEnter autocmd.

Also cleanup the other autocmds and put them in an augroup, so that they
re-sourcing the vimrc doesn't duplicate the autocmds. The Perl autocmd
has been moved to an ftplugin file.